+++ packages/cds-migrator/src/migrations/q42023/index.ts
@@ -5,11 +5,15 @@ import catchAllPropMigrations from './catchAllPropMigrations';
 import componentMigrations from './componentMigrations';
 import manualPropMigrations from './manualPropMigrations';
 import propMigrations from './propMigrations';
+import removedComponents from './removedComponents';
+import removedPropMigrations from './removedPropMigrations';
 
 export default async function main(tree: Tree) {
   await componentMigrations(tree);
+  await removedComponents(tree);
   await catchAllPropMigrations(tree);
   await propMigrations(tree);
+  await removedPropMigrations(tree);
   await manualPropMigrations(tree);
   await attributeValueToBooleanMigrations(tree);
 }