+++ packages/cds-migrator/src/migrations/update-5-0-0/removedImports.ts
@@ -11,22 +11,23 @@ const callback = (args: CreateMigrationParams) => {
   const { path, sourceFile } = args;
   // Check if the file contains any of the import statements
   const declarations = sourceFile.getImportDeclarations();
-  const deprecatedPaths = tokenMigrations.map(({ path: depPath }) => depPath);
+  const deprecatedPaths = tokenMigrations.map(({ path: deprecatedPath }) => deprecatedPath);
 
   if (!checkFileIncludesImport(sourceFile, deprecatedPaths)) return;
 
   declarations.forEach((declaration) => {
     const declarationPath = declaration.getModuleSpecifierValue();
-    const migrationConfig = tokenMigrations.find(({ path: depPath }) =>
-      declarationPath.includes(depPath),
+    const migrationConfig = tokenMigrations.find(({ path: deprecatedPath }) =>
+      declarationPath.includes(deprecatedPath),
     );
 
     if (migrationConfig) {
       const { warning, path: deprecatedPath } = migrationConfig;
       const title = `All exports from ${deprecatedPath} have been removed from CDS libraries`;
       const bodyLines = `You will need to manually migrate to an alternative. ${warning}`;
-      const outputMessage = `All manual migrations are output to the root of the repo in a file called:`;
-      const outputPath = `${output.underline(`${args.tree.root}/cds-migrator-output.md`)}`;
+      const outputMessage =
+        'All manual migrations are output to the root of the repo in a file called:';
+      const outputPath = output.underline(`${args.tree.root}/cds-migrator-output.md`);
       logWarning('Manual migration required!', [
         title,
         bodyLines,