+++ packages/cds-migrator/src/migrations/update-5-0-0/removedParams.ts
@@ -21,8 +21,9 @@ const callback = (args: CreateMigrationParams) => {
           if (hasRemovedParam) {
             const title = `The ${param} parameter has been removed from the ${name} function`;
             const bodyLines = `You will need to manually migrate to an alternative.`;
-            const outputMessage = `All manual migrations are output to the root of the repo in a file called:`;
-            const outputPath = `${output.underline(`${args.tree.root}/cds-migrator-output.md`)}`;
+            const outputMessage =
+              'All manual migrations are output to the root of the repo in a file called:';
+            const outputPath = output.underline(`${args.tree.root}/cds-migrator-output.md`);
             logWarning('Manual migration required!', [
               title,
               bodyLines,