+++ packages/common/system/FeatureFlagContext.ts
@@ -6,31 +6,37 @@ export const defaultFeatureFlags = {
   /**
    * Conventiently toggle all the Frontier flags at once.
    * @default false
+   * @deprecated Components will be updated in Q42023 to use frontier by default. This flag will be removed.
    */
   frontier: false,
   /**
    * Reduce Display2 fontSize/lineHeight. Add Display3 and Title4 components.
    * @default false
+   * @deprecated Components will be updated in Q42023 to use frontier by default. This flag will be removed.
    */
   frontierTypography: false,
   /**
    * Replace rounded border radius with pill style.
    * @default false
+   * @deprecated Components will be updated in Q42023 to use frontier by default. This flag will be removed.
    */
   frontierButton: false,
   /**
    * Change secondary palette from gray0 to gray60 at 10%.
    * @default false
+   * @deprecated Components will be updated in Q42023 to use frontier by default. This flag will be removed.
    */
   frontierColor: false,
   /**
    * Remove rounded border radius and elevation.
    * @default false
+   * @deprecated Components will be updated in Q42023 to use frontier by default. This flag will be removed.
    */
   frontierCard: false,
   /**
    * Dotted sparkline fills
    * @default false
+   * @deprecated Components will be updated in Q42023 to use frontier by default. This flag will be removed.
    */
   frontierSparkline: false,
   /** Web only - Leverages the CSS flex 'gap' property instead of inserting extra DOM nodes.